Refactor UI code

+++ hshchk-lib/src/block_hasher.rs
@@ -3,7 +3,7 @@ use crossbeam::crossbeam_channel::Sender;
 use std::sync::Arc;
 
 pub struct HashProgress {
-    pub bytes_processed: usize,
+    pub bytes_processed: u64,
 }
 
 pub trait BlockHasher {
@@ -14,15 +14,15 @@ pub trait BlockHasher {
     fn set_bytes_processed_event_sender_with_bytes_processed_notification_block_size(
         &mut self,
         sender: Sender<HashProgress>,
-        bytes_processed_notification_block_size: usize,
+        bytes_processed_notification_block_size: u64,
     );
-    fn bytes_processed_notification_block_size(&self) -> usize;
+    fn bytes_processed_notification_block_size(&self) -> u64;
     fn is_bytes_processed_event_sender_defined(&self) -> bool;
     fn handle_bytes_processed_event(&self, args: HashProgress);
     fn compute(&mut self, cancellation_token: Arc<CancellationToken>) {
         let mut bytes_read;
-        let mut running_notification_block_size = 0usize;
-        let mut bytes_processed = 0usize;
+        let mut running_notification_block_size = 0u64;
+        let mut bytes_processed = 0u64;
         let bytes_processed_notification_block_size =
             self.bytes_processed_notification_block_size();
         loop {
@@ -36,8 +36,8 @@ pub trait BlockHasher {
                 if self.is_bytes_processed_event_sender_defined()
                     && bytes_processed_notification_block_size > 0
                 {
-                    bytes_processed += bytes_read;
-                    running_notification_block_size += bytes_read;
+                    bytes_processed += bytes_read as u64;
+                    running_notification_block_size += bytes_read as u64;
                     if running_notification_block_size >= bytes_processed_notification_block_size
                         || bytes_read == 0
                     {
